Backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.17
